ASG works

+++ lib/configure.sh
@@ -15,6 +15,7 @@ docker run hello-world
 
 # pull drugseqr
 docker pull alexvpickering/drugseqr
+docker pull openanalytics/shinyproxy-demo
 docker network create sp-example-net
 
 # get Dockerfile for ShinyProxy image
@@ -27,50 +28,50 @@ wget https://raw.githubusercontent.com/hms-dbmi/drugseqr.sp/master/Dockerfile
 wget https://raw.githubusercontent.com/hms-dbmi/drugseqr.sp/master/application.yml
 wget https://raw.githubusercontent.com/hms-dbmi/drugseqr.sp/master/app.html
 wget https://raw.githubusercontent.com/hms-dbmi/drugseqr.sp/master/login.html
+wget https://raw.githubusercontent.com/hms-dbmi/drugseqr.sp/master/index.html
 
-# modify application.yml to use cognito
-# replacement variables have to be set e.g. in user data
-if ["$USE_COGNITO" = true]; then
-  sed -i "s/# /" application.yml # uncomment single comments
-  sed -i "s/authentication: none/authentication: openid" application.yml
+# replacement variables set in user data of DseqrAsgStack
+if [ "$USE_COGNITO" = true ]
+then
+  # modify application.yml to use cognito and serve shinyproxy on port 80
+  SHINYPROXY_PORT=80
+  sed -i "s/authentication: none/authentication: openid/" application.yml
   sed -i "s/{region}/$REGION/" application.yml
-  sed -i "s/{userPoolId}/$USER_POOL_ID/" application.yml
-  sed -i "s/{cognito_domain_prefix}/$COGNITO_DOMAIN/" application.yml
+  sed -i "s/{user_pool_id}/$USER_POOL_ID/" application.yml
+  sed -i "s/{auth_domain}/$AUTH_DOMAIN/" application.yml
   sed -i "s/{client_id}/$CLIENT_ID/" application.yml
-  sed -i "s/{your_host_url}/$HOST_URL/" application.yml
+  sed -i "s/{host_url}/$HOST_URL/" application.yml
   sed -i "s/client-id:/client-id: $CLIENT_ID/" application.yml
   sed -i "s/client-secret:/client-secret: $CLIENT_SECRET/" application.yml
+else
+  # serve shinyproxy behind nginx
+  SHINYPROXY_PORT=8080
+  # install nginx
+  apt install nginx -y
+  ufw allow ssh
+  ufw allow 'Nginx Full'
+  ufw --force enable
+  # setup nginx
+  cd /etc/nginx/sites-available
+  wget -O $HOST_URL https://raw.githubusercontent.com/hms-dbmi/drugseqr.sp/master/nginx.conf
+  # change server name
+  sed -i "s/localhost/$HOST_URL www.$HOST_URL/" $HOST_URL
+  # enable site and restart nginx
+  ln -s /etc/nginx/sites-available/$HOST_URL /etc/nginx/sites-enabled/
+  rm /etc/nginx/sites-enabled/default
+  systemctl restart nginx
 fi
 
 docker build -t drugseqr.sp .
 
-# install nginx
-apt install nginx -y
-ufw allow ssh
-ufw allow 'Nginx Full'
-ufw --force enable
-
-# setup nginx
-cd /etc/nginx/sites-available
-wget -O drugseqr.com https://raw.githubusercontent.com/hms-dbmi/drugseqr.sp/master/nginx.conf
-
-# if e.g. launching at drugseqr.com
-sed -i 's/localhost/drugseqr.com www.drugseqr.com/' drugseqr.com
-
-# if e.g. launching with sjia data
-# sed -i 's/example/sjia/' drugseqr.com
-
-ln -s /etc/nginx/sites-available/drugseqr.com /etc/nginx/sites-enabled/
-rm /etc/nginx/sites-enabled/default
-systemctl restart nginx
-
 # setup ssl
-# apt-get install certbot -y
-# apt-get install python-certbot-nginx -y
-# certbot --non-interactive --nginx \
-#  -d drugseqr.com -d www.drugseqr.com \
-#  --agree-tos -m alexvpickering@gmail.com
-
+if [ "$GET_CERT" = true ]; then
+  apt-get install certbot -y
+  apt-get install python-certbot-nginx -y
+  certbot --non-interactive --nginx \
+  -d $HOST_URL -d www.$HOST_URL \
+  --agree-tos -m alexvpickering@gmail.com
+fi
 
 # init example app
 docker run --rm \
@@ -78,18 +79,17 @@ docker run --rm \
   alexvpickering/drugseqr R -e "drugseqr::init_drugseqr('example')"
 
 # get data for example app
-cd /srv/drugseqr/
-rm -rf example
-if [ ! -f example_data.tar.gz ]; then
+if [ "$EXAMPLE_DATA" = true ] && [ ! -f "example_data.tar.gz" ]; then
+  cd /srv/drugseqr/
+  rm -rf example
   wget https://drugseqr.s3.us-east-2.amazonaws.com/example_data.tar.gz
   tar -xzvf example_data.tar.gz
   rm example_data.tar.gz
 fi
 
-
 # run app
 docker run -d --restart always \
  -v /var/run/docker.sock:/var/run/docker.sock \
  --net sp-example-net \
-  -p 8080:8080 \
+  -p $SHINYPROXY_PORT:8080 \
    drugseqr.sp
